[TSK-276] refactor: update footer and login templates to use namespaced URL routes, improve error handling in route registration
