+++ app/templates/handler/index.html
@@ -1,58 +1,57 @@
 <!DOCTYPE html>
 <html lang="en">
-    <head>
-        <meta charset="utf-8" />
-        <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-        <meta name="viewport"
-              content="width=device-width, initial-scale=1, shrink-to-fit=no" />
-        <meta name="description" content="" />
-        <meta name="keywords"
-              content="error, GuardEPI, dashboard, privacy policy, terms and conditions" />
-        <meta name="author" content="" />
-        <title>Erro {{ code }} - GuardEPI</title>
-        <link href="{{ url_for('static', filename = 'css/styles.css') }}"
-              rel="stylesheet" />
-        <script src="https://use.fontawesome.com/releases/v6.3.0/js/all.js"
-                crossorigin="anonymous"></script>
-    </head>
-    <body>
-        <div id="layoutError">
-            <div id="layoutError_content">
-                <main>
-                    <div class="container">
-                        <div class="row justify-content-center">
-                            <div class="col-lg-6">
-                                <div class="text-center mt-4">
-                                    <h1 class="display-1">{{ code }}</h1>
-                                    <p class="lead">{{ name }}</p>
-                                    <p>{{ desc }}</p>
-                                    <a href="{{ url_for('dash.dashboard') }}">
-                                        <i class="fas fa-arrow-left me-1"></i>
-                                        Retornar para a Dashboard
-                                    </a>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                </main>
+
+<head>
+  <meta charset="utf-8" />
+  <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+  <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
+  <meta name="description" content="" />
+  <meta name="keywords" content="error, GuardEPI, dashboard, privacy policy, terms and conditions" />
+  <meta name="author" content="" />
+  <title>Erro {{ code }} - GuardEPI</title>
+  <link href="{{ url_for('static', filename = 'css/styles.css') }}" rel="stylesheet" />
+  <script src="https://use.fontawesome.com/releases/v6.3.0/js/all.js" crossorigin="anonymous"></script>
+</head>
+
+<body>
+  <div id="layoutError">
+    <div id="layoutError_content">
+      <main>
+        <div class="container">
+          <div class="row justify-content-center">
+            <div class="col-lg-6">
+              <div class="text-center mt-4">
+                <h1 class="display-1">{{ code }}</h1>
+                <p class="lead">{{ name }}</p>
+                <p>{{ desc }}</p>
+                <a href="{{ url_for('dash.dashboard') }}">
+                  <i class="fas fa-arrow-left me-1"></i>
+                  Retornar para a Dashboard
+                </a>
+              </div>
             </div>
-            <div id="layoutError_footer">
-                <footer class="py-4 bg-light mt-auto">
-                    <div class="container-fluid px-4">
-                        <div class="d-flex align-items-center justify-content-between small">
-                            <div class="text-muted">Copyright © GuardEPI 2024</div>
-                            <div>
-                                <a href="{{ url_for('politica_privacidade') }}">Política de Privacidade</a>
-                                ·
-                                <a href="{{ url_for('termos_uso') }}">Termos &amp; Condições</a>
-                            </div>
-                        </div>
-                    </div>
-                </footer>
-            </div>
+          </div>
         </div>
-        <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/js/bootstrap.bundle.min.js"
-                crossorigin="anonymous"></script>
-        <script src="{{ url_for('static', filename = 'js/scripts.js') }}"></script>
-    </body>
+      </main>
+    </div>
+    <div id="layoutError_footer">
+      <footer class="py-4 bg-light mt-auto">
+        <div class="container-fluid px-4">
+          <div class="d-flex align-items-center justify-content-between small">
+            <div class="text-muted">Copyright © GuardEPI 2024</div>
+            <div>
+              <a href="{{ url_for('index.politica_privacidade') }}">Política de Privacidade</a>
+              ·
+              <a href="{{ url_for('index.termos_uso') }}">Termos &amp; Condições</a>
+            </div>
+          </div>
+        </div>
+      </footer>
+    </div>
+  </div>
+  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/js/bootstrap.bundle.min.js"
+    crossorigin="anonymous"></script>
+  <script src="{{ url_for('static', filename = 'js/scripts.js') }}"></script>
+</body>
+
 </html>
