+++ app/templates/login.html
@@ -75,9 +75,9 @@
                     <div class="d-flex align-items-center justify-content-between small">
                         <div class="text-muted">Copyright © GuardEPI 2024</div>
                         <div>
-                            <a href="{{ url_for('politica_privacidade') }}">Política de Privacidade</a>
+                            <a href="{{ url_for('index.politica_privacidade') }}">Política de Privacidade</a>
                             ·
-                            <a href="{{ url_for('termos_uso') }}">Termos &amp; Condições</a>
+                            <a href="{{ url_for('index.termos_uso') }}">Termos &amp; Condições</a>
                         </div>
                     </div>
                 </div>
